UpdaterFixes and Def changes - Version 1.0.0.3
fixed a bug where the software would be unable to correctly see it's own version.

Added animstudio to the def file back.
Removed flver editor from Demon's Souls thanks to lenox for pointing out the incompatibility.

Added an updater for the updater, in order to allow pushing new code to it.

now HoodieUpdater is shipped in a zip with the installer, instead of sparse in a folder.  It gets installed at first launch.